+++ ldm/dream/generator/__init__.py
@@ -1,0 +1,4 @@
+'''
+Initialization file for the ldm.dream.generator package
+'''
+from .base import Generator
